% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBAM}
\alias{exampleBAM}
\title{An example BAM file.}
\format{
A list with 7 factors, each contains 184 variables:
\describe{
\item{rname}{chromosome of the read}
\item{qname}{read ID list}
\item{seq}{sequence of the read, in DNAString}
\item{strand}{strand of the read}
\item{cigar}{CIGAR sequence of the read}
\item{qual}{Phred quality of the read}
\item{pos}{starting position of the read}
...
}
}
\usage{
exampleBAM
}
\description{
A BAM file containing the information of eight mutations.
}
\keyword{datasets}
