% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMutCall}
\alias{exampleMutCall}
\title{An example mutated read ID file.}
\format{
A list with 7 factors, each contains 184 variables:
\describe{
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{Mut_ID}{mutated read ID}
\item{Mut}{mutation detail}
...
}
}
\usage{
exampleMutCall
}
\description{
A dataset containing the information of mutated read.
}
\keyword{datasets}
