% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hairpin_check.R
\name{fun_hairpin_check}
\alias{fun_hairpin_check}
\title{Hairpin-structure sequence check function}
\usage{
fun_hairpin_check(Hairpin_seq_tmp, Ref_seq, Hairpin_length, hair)
}
\arguments{
\item{Hairpin_seq_tmp}{The sequence to be checked.}

\item{Ref_seq}{Reference sequence around the mutation.}

\item{Hairpin_length}{The temporal length of hairpin sequences.}

\item{hair}{The length of sequences to be checked.}
}
\value{
list(Hairpin_length, whether hairpin sequences exist or not)
}
\description{
This function attempts to find hairpin structure sequences.
}
\examples{
fun_hairpin_check(DNAString("GGGGGGGGAAAAAAAA"), DNAString("GGGGGGGGAAAAAAAATATATATATTTTTTTTCCCCCCCC"), 15, 16)
}
