% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hairpin_trimming.R
\name{fun_hairpin_trimming}
\alias{fun_hairpin_trimming}
\title{Adapter sequence trimming function}
\usage{
fun_hairpin_trimming(
  Hairpin_seq,
  mut_read_strand,
  ADAPTER_SEQ_1,
  ADAPTER_SEQ_2
)
}
\arguments{
\item{Hairpin_seq}{The sequence to be trimmed.}

\item{mut_read_strand}{The strand of the sequence, "+" or "-".}

\item{ADAPTER_SEQ_1}{The Read 1 adapter sequence of the library.}

\item{ADAPTER_SEQ_2}{The Read 2 adapter sequence of the library.}
}
\value{
Adapter-trimmed Hairpin_seq
}
\description{
This function attempts to find and cut the adapter sequences in the ends.
}
\examples{
fun_hairpin_trimming(DNAString("GGAAAAAAATCTCTCAACA"), "+", "AGATCC")
}
