% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_homology.R
\name{fun_homology}
\alias{fun_homology}
\title{Homology check function.}
\usage{
fun_homology(MSEC, df_distant, Minimum_Homology_search_length, PROGRESS_BAR)
}
\arguments{
\item{MSEC}{Mutation filtering information.}

\item{df_distant}{Sequences to be checked.}

\item{Minimum_Homology_search_length}{Minimum length to define "homologous".}

\item{PROGRESS_BAR}{"Y": You can see the progress visually.}
}
\value{
MSEC
}
\description{
This function attempts to search the homologous regions.
}
