% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_insert_length.R
\name{fun_insert_length}
\alias{fun_insert_length}
\title{Insert length checker for BAM file}
\usage{
fun_insert_length(bam_file)
}
\arguments{
\item{bam_file}{Path of the BAM file.}
}
\value{
df_bam
}
\description{
This function attempts to check the insert length in BAM file.
}
\examples{
fun_insert_length(
  system.file("extdata", "test.bam", package = "MicroSEC")
)
}
