% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_BAM.R
\name{fun_load_BAM}
\alias{fun_load_BAM}
\title{BAM file loader}
\usage{
fun_load_BAM(BAM_FILE)
}
\arguments{
\item{BAM_FILE}{Path of the BAM file.}
}
\value{
df_BAM
}
\description{
This function attempts to load the BAM file.
}
\examples{
fun_load_BAM("./source/Cell_line/PC9_Cell_line_Ag_TDv4.realigned.bam")
}
