% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_chr_no.R
\name{fun_load_chr_no}
\alias{fun_load_chr_no}
\title{Chromosome number loading function.}
\usage{
fun_load_chr_no(GENOME)
}
\arguments{
\item{GENOME}{Human or Mouse genome.}
}
\value{
Chr_No
}
\description{
This function attempts to load the chromosome number.
}
\examples{
fun_load_chr_no("Human")
}
