% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_genome.R
\name{fun_load_genome}
\alias{fun_load_genome}
\title{Genome loading function.}
\usage{
fun_load_genome(organism)
}
\arguments{
\item{organism}{Human or Mouse genome.}
}
\value{
ref_genome
}
\description{
This function attempts to load the appropriate genome.
}
\examples{
fun_load_genome("Human")
}
