% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_ID.R
\name{fun_load_ID}
\alias{fun_load_ID}
\title{Mutation-supporting read ID files loader}
\usage{
fun_load_ID(MUTATION_SUPPORTING_READ_LIST)
}
\arguments{
\item{MUTATION_SUPPORTING_READ_LIST}{Path of the read ID information directory.}
}
\value{
df_mut_call
}
\description{
This function attempts to load the read ID information files.
}
\examples{
fun_load_ID("./source/PC9_Cell_line")
}
