% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_normal_support.R
\name{fun_normal_support}
\alias{fun_normal_support}
\title{Supporting length calculation function.}
\usage{
fun_normal_support(df_cigar, df_pos, mut_position, alt_length, indel_status)
}
\arguments{
\item{df_cigar}{The CIGAR data of the read.}

\item{df_pos}{The start position of the read.}

\item{mut_position}{The mutation position in the genome.}

\item{alt_length}{The length of altered bases.}

\item{indel_status}{The mutation is indel or not.}
}
\value{
list(pre_support_length, post_support_length, short_support_length)
}
\description{
This function attempts to calculate supporting lengths of a read.
}
