% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_repeat_check.R
\name{fun_repeat_check}
\alias{fun_repeat_check}
\title{Repeat check function.}
\usage{
fun_repeat_check(Rep_A, Rep_B, Ref_seq, Width, Del)
}
\arguments{
\item{Rep_A}{The shorter sequence of \link{Ref} and \link{Alt}.}

\item{Rep_B}{The longer sequence of \link{Ref} and \link{Alt}.}

\item{Ref_seq}{Reference sequence around the mutation.}

\item{Width}{Search length for Ref_seq.}

\item{Del}{Insertion: 0, Deletion: 1}
}
\value{
list(Pre_rep_status, Post_rep_status, Homopolymer_status)
}
\description{
This function attempts to check the repetitive sequence around the mutation.
}
\examples{
fun_repeat_check(DNAString("A"), DNAString("AATC"), DNAString("GGAAAAAAATCTCTCAACA"), 5, 0)
}
