% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save.R
\name{fun_save}
\alias{fun_save}
\title{Save function.}
\usage{
fun_save(MicroSEC, wd)
}
\arguments{
\item{MicroSEC}{Mutation filtering information.}

\item{wd}{The directory to save.}
}
\description{
This function attempts to save the filtering results.
}
