% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save_gz.R
\name{fun_save_gz}
\alias{fun_save_gz}
\title{Save function.}
\usage{
fun_save_gz(MSEC, OUTPUT)
}
\arguments{
\item{MSEC}{Mutation filtering information.}

\item{OUTPUT}{Output file name (full path).}
}
\description{
This function attempts to save the filtering results.
}
