% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_setting.R
\name{fun_setting}
\alias{fun_setting}
\title{Mutated position search function.}
\usage{
fun_setting(PRE, POST, Neighbor_seq, neighbor_length, Alt_length)
}
\arguments{
\item{PRE}{The 5' side bases of the sequence for searching.}

\item{POST}{The 3' side bases of the sequence for searching.}

\item{Neighbor_seq}{Short reference sequence around the mutation.}

\item{neighbor_length}{The length from the mutation to the ends of the short reference sequence.}

\item{Alt_length}{The length of altered bases.}
}
\value{
list(Pre_search_length, Post_search_length, Peri_seq_1, Peri_seq_2)
}
\description{
This function attempts to find the mutated bases in each read.
}
\examples{
fun_setting(0, 20, "AGATCCT", 3, 1)
}
