% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{fun_summary}
\alias{fun_summary}
\title{Summarizing function.}
\usage{
fun_summary(MSEC)
}
\arguments{
\item{MSEC}{Mutation filtering information.}
}
\value{
MSEC
}
\description{
This function summarizes the filtering results.
}
