% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_support.R
\name{fun_support}
\alias{fun_support}
\title{Supporting length calculation function.}
\usage{
fun_support(
  df_cigar,
  df_seq,
  mut_read_strand,
  ADAPTOR_SEQ,
  mut_position,
  Alt_length,
  indel_status
)
}
\arguments{
\item{df_cigar}{The CIGAR data of the read.}

\item{df_seq}{The sequence to be checked.}

\item{mut_read_strand}{The strand of the sequence, "+" or "-".}

\item{ADAPTOR_SEQ}{The adapter sequence of the library.}

\item{mut_position}{The mutation position in the read.}

\item{Alt_length}{The length of altered bases.}
}
\value{
list(Pre_support_length, Post_support_length, Soft_Clipped_read)
}
\description{
This function attempts to calculate supporting lengths of a read.
}
