% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_zero.R
\name{fun_zero}
\alias{fun_zero}
\title{Divide function without 0/0 errors}
\usage{
fun_zero(a, b)
}
\arguments{
\item{a, b}{Integers}
}
\value{
a divided by b
}
\description{
This function attempts to divide without 0/0 errors.
}
