% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homology_searched}
\alias{homology_searched}
\title{An example sequence infromation file.}
\format{
A list with 7 factors, each contains 460 variables:
\describe{
\item{sample_name}{sample name}
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{Direction}{5' (pre) or 3' (post) sequence of the mutated base}
\item{Seq}{sequence for homology search}
...
}
}
\usage{
homology_searched
}
\description{
A dataset containing the information of reads for homology search.
}
\keyword{datasets}
