% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroSEC-package.R
\docType{package}
\name{MicroSEC-package}
\alias{MicroSEC}
\alias{MicroSEC-package}
\title{MicroSEC: MicroSEC sequence error filterinf pipeline for FFPE samples}
\description{
Clinical sequencing of tumor is usually performed on formalin-fixed and paraffin-embedded (FFPE) samples and have many sequencing errors. We found that the majority of these errors are detected in chimeric reads caused by single-strand DNA with microhomology. Our filtering pipeline, MicroSEC, focuses on the uneven distribution of the artifacts in each read and removes such errors in FFPE samples without over-eliminating the true mutations detected in fresh frozen samples.
}
\author{
\strong{Maintainer}: Masachika Ikegami \email{ikegamitky@gmail.com}

}
\keyword{internal}
