% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msec_analyzed}
\alias{msec_analyzed}
\title{An example mutation file.}
\format{
A list with 37 factors, each contains 29 variables
\describe{
\item{Sample}{sample name}
\item{Mut_type}{mutation type}
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{SimpleRepeat_TRF}{mutation locating repeat sequence}
\item{Neighborhood_sequence}{neighborhood sequence}
\item{read_length}{read length}
\item{total_read}{number of mutation supporting reads}
\item{soft_clipped_read}{number of soft-clipped reads}
\item{flag_hairpin}{number of reads produced by hairpin structure}
\item{pre_support_length}{maximum 5'-supporting length}
\item{post_support_length}{maximum 3'-supporting length}
\item{short_support_length}{minimum supporting length}
\item{pre_farthest}{5'-farthest supported base from the mutated base}
\item{post_farthest}{3'-farthest supported base from the mutated base}
\item{low_quality_base_rate_under_q18}{low quality base rate}
\item{low_quality_pre}{low quality base rate of 5'- side}
\item{low_quality_post}{low quality base rate of 3'- side}
\item{distant_homology_rate}{rate of reads derived from homologous regions}
\item{soft_clipped_rate}{rate of soft clipped reads}
\item{prob_filter_1}{possibility of short-supporting length}
\item{prob_filter_3_pre}{possibility of 5'-supporting length}
\item{prob_filter_3_post}{possibility of 3'-supporting length}
\item{filter_1_mutation_intra_hairpin_loop}{filter 1}
\item{filter_2_hairpin_structure}{filter 2}
\item{filter_3_microhomology_induced_mutation}{filter 3}
\item{filter_4_highly_homologous_region}{filter 4}
\item{filter_5_soft_clipping}{filter 5}
\item{filter_6_simple_repeat}{filter 6}
\item{filter_7_mutation_at_homopolymer}{filter 7}
\item{filter_8_low_quality}{filter 8}
\item{msec_filter_123}{any of filter 1-3}
\item{msec_filter_1234}{any of filter 1-4}
\item{msec_filter_all}{any of filter 1-8}
\item{comment}{comment}
...
}
}
\usage{
msec_analyzed
}
\description{
A dataset containing the information of eight mutations processed by the
fun_homology function.
}
\keyword{datasets}
