% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msec_summarized}
\alias{msec_summarized}
\title{An example mutation file.}
\format{
A list with 53 variables:
\describe{
\item{Sample}{sample name}
\item{Gene}{altered gene}
\item{HGVS.c}{base change}
\item{HGVS.p}{protein change}
\item{Mut_type}{mutation type}
\item{Total_QV>=20}{total reads}
\item{\%Alt}{altered base ratio}
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{SimpleRepeat_TRF}{mutation locating repeat sequence}
\item{Neighborhood_sequence}{neighborhood sequence}
\item{Transition}{base change type}
\item{read_length}{read length}
\item{mut_type}{mutation type}
\item{alt_length}{length of the mutated bases}
\item{total_read}{number of mutation supporting reads}
\item{soft_clipped_read}{number of soft-clipped reads}
\item{flag_hairpin}{number of reads produced by hairpin structure}
\item{hairpin_length}{maximum length of palindromic sequences}
\item{pre_support_length}{muximum 5'-supporting length}
\item{post_support_length}{muximum 3'-supporting length}
\item{short_support_length}{minimum supporting length}
\item{pre_minimum_length}{minimum 5'-supporting length}
\item{post_minimum_length}{minimum 3'-supporting length}
\item{low_quality_base_rate_under_q18}{low quality base rate}
\item{pre_rep_status}{5'-repeat sequence length}
\item{post_rep_status}{3'-repeat sequence length}
\item{homopolymer_status}{homopolymer sequence length}
\item{indel_status}{whether the mutation is indel or not}
\item{indel_length}{length of indel mutation}
\item{distant_homology}{number of reads derived from homologous regions}
\item{penalty_pre}{5'-penalty score by the mapper}
\item{penalty_post}{3'-penalty score by the mapper}
\item{caution}{comment}
\item{distant_homology_rate}{rate of reads derived from homologous regions}
\item{pre_minimum_length_adj}{adjusted pre_minimum_length}
\item{post_minimum_length_adj}{adjusted pre_minimum_length}
\item{pre_support_length_adj}{adjusted pre_minimum_length}
\item{post_support_length_adj}{adjusted pre_minimum_length}
\item{shortest_support_length_adj}{the shortest short_support_length}
\item{minimum_length_1}{theoretically minimum 5'-supporting length}
\item{minimum_length_2}{theoretically minimum 3'-supporting length}
\item{minimum_length}{theoretically minimum supporting length}
\item{short_support_length_adj}{adjusted short_support_length}
\item{altered_length}{substituted/inserted length}
\item{half_length}{half of the read length}
\item{short_support_length_total}{range of short_support_length}
\item{pre_support_length_total}{range of pre_support_length}
\item{post_support_length_total}{range of post_support_length}
\item{half_length_total}{range of possible short_support_length}
\item{total_length_total}{range of possible supporting length}
...
}
}
\usage{
msec_summarized
}
\description{
A dataset containing the information of eight mutations processed by the
fun_homology function.
}
\keyword{datasets}
