% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_depth_checked}
\alias{mut_depth_checked}
\title{An example sequence infromation file.}
\format{
A list with 7 factors, each contains 460 variables:
\describe{
\item{Zero}{zero vector}
\item{Depth0}{number of reads whose 5'-supporting length is <= 0}
\item{Depth1}{number of reads whose 5'-supporting length is <= 1}
\item{Depth2}{number of reads whose 5'-supporting length is <= 2}
\item{Depth3}{number of reads whose 5'-supporting length is <= 3}
...
}
}
\usage{
mut_depth_checked
}
\description{
A dataset containing the information of reads for homology search.
}
\keyword{datasets}
