% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBam}
\alias{exampleBam}
\title{An example BAM file.}
\format{
A list with 8 factors, each contains 46527 variables:
\describe{
\item{rname}{chromosome of the read}
\item{qname}{read ID list}
\item{seq}{sequence of the read, in DNAString}
\item{strand}{strand of the read}
\item{cigar}{CIGAR sequence of the read}
\item{qual}{Phred quality of the read}
\item{pos}{starting position of the read}
\item{isize}{insert size of the read}
...
}
}
\usage{
exampleBam
}
\description{
A BAM file containing the information of eight mutations.
}
\keyword{datasets}
