% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMutation}
\alias{exampleMutation}
\title{An example mutation file.}
\format{
A list with 8 factors, each contains 29 variables
\describe{
\item{Sample}{sample name}
\item{Mut_type}{mutation type}
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{SimpleRepeat_TRF}{mutation locating repeat sequence}
\item{Neighborhood_sequence}{neighborhood sequence}
...
}
}
\usage{
exampleMutation
}
\description{
A dataset containing the information of eight mutations.
}
\keyword{datasets}
