% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analysis.R
\name{fun_analysis}
\alias{fun_analysis}
\title{Analyzing function.}
\usage{
fun_analysis(
  MSEC,
  Mut_depth,
  threshold_p = 10^(-6),
  threshold_hairpin_ratio = 0.5,
  threshold_hairpin_length = 30,
  threshold_soft_clip_ratio = 0.9,
  threshold_short_length = 0.8,
  threshold_distant_homology = 0.2,
  threshold_low_quality_rate = 0.1,
  Homopolymer_length = 15
)
}
\arguments{
\item{MSEC}{Mutation filtering information.}
}
\value{
MSEC
}
\description{
This function analyzes the filtering results.
}
