% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert.R
\name{fun_convert}
\alias{fun_convert}
\title{Mutation data file converter}
\usage{
fun_convert(mutation_file, organism)
}
\arguments{
\item{mutation_file}{Path of the mutation information file.}

\item{organism}{Human or Mouse genome.}
}
\value{
df_mutation
}
\description{
This function attempts to convert the mutation information file.
}
\examples{
\donttest{
fun_convert(
  system.file("extdata", "convert_test.xlsx", package = "MicroSEC"),
  "hg19"
)
}
}
