% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hairpin_check.R
\name{fun_hairpin_check}
\alias{fun_hairpin_check}
\title{Hairpin-structure sequence check function}
\usage{
fun_hairpin_check(hairpin_seq_tmp, ref_seq, hairpin_length, hair)
}
\arguments{
\item{hairpin_seq_tmp}{The sequence to be checked.}

\item{ref_seq}{Reference sequence around the mutation.}

\item{hairpin_length}{The temporal length of hairpin sequences.}

\item{hair}{The length of sequences to be checked.}
}
\value{
list(hairpin_length, whether hairpin sequences exist or not)
}
\description{
This function attempts to find hairpin structure sequences.
}
