% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hairpin_trimming.R
\name{fun_hairpin_trimming}
\alias{fun_hairpin_trimming}
\title{Adapter sequence trimming function}
\usage{
fun_hairpin_trimming(hairpin_seq, mut_read_strand, adapter_1, adapter_2)
}
\arguments{
\item{hairpin_seq}{The sequence to be trimmed.}

\item{mut_read_strand}{The strand of the sequence, "+" or "-".}

\item{adapter_1}{The Read 1 adapter sequence of the library.}

\item{adapter_2}{The Read 2 adapter sequence of the library.}
}
\value{
Adapter-trimmed hairpin_seq
}
\description{
This function attempts to find and cut the adapter sequences in the ends.
}
