% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_id.R
\name{fun_load_id}
\alias{fun_load_id}
\title{Mutation-supporting read ID files loader}
\usage{
fun_load_id(read_list)
}
\arguments{
\item{read_list}{Path of the read ID information
directory.}
}
\value{
df_mut_call
}
\description{
This function attempts to load the read ID information files.
}
\examples{
fun_load_id(
  system.file("extdata", package = "MicroSEC")
)
}
