% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_mutation.R
\name{fun_load_mutation}
\alias{fun_load_mutation}
\title{Mutation data file loader}
\usage{
fun_load_mutation(mutation_file, sample_name)
}
\arguments{
\item{mutation_file}{Path of the mutation information file.}

\item{sample_name}{Sample name.}
}
\value{
df_mutation
}
\description{
This function attempts to load the mutation information file.
}
\examples{
\dontrun{
fun_load_mutation("./source/CCLE.xlsx", "PC9")
}
}
