% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_mutation_gz.R
\name{fun_load_mutation_gz}
\alias{fun_load_mutation_gz}
\title{Mutation data file loader}
\usage{
fun_load_mutation_gz(mutation_file)
}
\arguments{
\item{mutation_file}{Path of the mutation information file.}
}
\value{
df_mutation
}
\description{
This function attempts to load the mutation information file.
}
\examples{
fun_load_mutation_gz(
  system.file("extdata", "test_mutation.tsv", package = "MicroSEC")
)
}
