% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_check.R
\name{fun_read_check}
\alias{fun_read_check}
\title{Read check function.}
\usage{
fun_read_check(
  df_mutation,
  df_BAM,
  df_mut_call,
  genome,
  Chr_No,
  SAMPLE_NAME,
  READ_length,
  ADAPTER_SEQ_1,
  ADAPTER_SEQ_2,
  Short_Homology_search_length,
  PROGRESS_BAR
)
}
\arguments{
\item{df_mutation}{Mutation information.}

\item{df_BAM}{Data from the BAM file.}

\item{df_mut_call}{Read ID list.}

\item{genome}{Reference genome for the data.}

\item{Chr_No}{Chromosome number.}

\item{SAMPLE_NAME}{Sample name (character)}

\item{READ_length}{The read length in the sequence.}

\item{ADAPTER_SEQ_1}{The Read 1 adapter sequence of the library.}

\item{ADAPTER_SEQ_2}{The Read 2 adapter sequence of the library.}

\item{Short_Homology_search_length}{Small sequence for homology search.}

\item{PROGRESS_BAR}{"Y": You can see the progress visually.}
}
\value{
list(MSEC, Homology_search)
}
\description{
This function attempts to check the mutation profile in each read.
}
