% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_repeat_check.R
\name{fun_repeat_check}
\alias{fun_repeat_check}
\title{Repeat check function.}
\usage{
fun_repeat_check(rep_a, rep_b, ref_seq, ref_width, del)
}
\arguments{
\item{rep_a}{The shorter sequence of Ref and Alt.}

\item{rep_b}{The longer sequence of Ref and Alt.}

\item{ref_seq}{Reference sequence around the mutation.}

\item{ref_width}{Search length for ref_seq.}

\item{del}{Insertion: 0, Deletion: 1}
}
\value{
list(pre_rep_status, post_rep_status, homopolymer_status)
}
\description{
This function attempts to check the repetitive sequence around the mutation.
}
