% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save.R
\name{fun_save}
\alias{fun_save}
\title{Save function.}
\usage{
fun_save(msec, output)
}
\arguments{
\item{msec}{Mutation filtering information.}

\item{output}{output file name (full path).}
}
\description{
This function attempts to save the filtering results.
}
\examples{
\donttest{
fun_save(msec_analyzed, "./test_data.tsv")
}
}
