% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_setting.R
\name{fun_setting}
\alias{fun_setting}
\title{Mutated position search function.}
\usage{
fun_setting(pre, post, neighbor_seq, neighbor_length, alt_length)
}
\arguments{
\item{pre}{The 5' side bases of the sequence for searching.}

\item{post}{The 3' side bases of the sequence for searching.}

\item{neighbor_seq}{Short reference sequence around the mutation.}

\item{neighbor_length}{The length from the mutation to the ends of the short
reference sequence.}

\item{alt_length}{The length of altered bases.}
}
\value{
list(pre_search_length, post_search_length, peri_seq_1, peri_seq_2)
}
\description{
This function attempts to find the mutated bases in each read.
}
