% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{fun_summary}
\alias{fun_summary}
\title{Summarizing function.}
\usage{
fun_summary(msec)
}
\arguments{
\item{msec}{Mutation filtering information.}
}
\value{
msec
}
\description{
This function summarizes the filtering results.
}
\examples{
fun_summary(msec_homology_searched)
}
