% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msec_analyzed}
\alias{msec_analyzed}
\title{An example mutation file.}
\format{
A list with 38 variables:
\describe{
\item{Sample}{sample name}
\item{Gene}{altered gene}
\item{HGVS.c}{base change}
\item{HGVS.p}{protein change}
\item{Mut_type}{mutation type}
\item{Total_QV>=20}{total reads}
\item{\%Alt}{altered base ratio}
\item{Chr}{altered chromosome}
\item{Pos}{altered position}
\item{Ref}{reference base}
\item{Alt}{altered base}
\item{SimpleRepeat_TRF}{mutation locating repeat sequence}
\item{Neighborhood_sequence}{neighborhood sequence}
\item{Transition}{base change type}
\item{read_length}{read length}
\item{total_read}{number of mutation supporting reads}
\item{soft_clipped_read}{number of soft-clipped reads}
\item{flag_hairpin}{number of reads produced by hairpin structure}
\item{pre_support_length}{muximum 5'-supporting length}
\item{post_support_length}{muximum 3'-supporting length}
\item{short_support_length}{minimum supporting length}
\item{low_quality_base_rate_under_q18}{low quality base rate}
\item{distant_homology_rate}{rate of reads derived from homologous regions}
\item{prob_filter_1}{possibility of short-supporting length}
\item{prob_filter_3_pre}{possibility of 5'-supporting length}
\item{prob_filter_3_post}{possibility of 3'-supporting length}
\item{filter_1_mutation_intra_hairpin_loop}{filter 1}
\item{filter_2_hairpin_structure}{filter 2}
\item{filter_3_microhomology_induced_mutation}{filter 3}
\item{filter_4_soft_clipping}{filter 4}
\item{filter_5_highly_homologous_region}{filter 5}
\item{filter_6_simple_repeat}{filter 6}
\item{filter_7_c_to_t_artifact}{filter 7}
\item{filter_8_mutation_at_homopolymer}{filter 8}
\item{msec_filter_1234}{any of filter 1-4}
\item{msec_filter_12345}{any of filter 1-5}
\item{msec_filter_all}{any of filter 1-8}
\item{comment}{comment}
...
}
}
\usage{
msec_analyzed
}
\description{
A dataset containing the information of eight mutations processed by the
fun_homology function.
}
\keyword{datasets}
