% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_depth_checked}
\alias{mut_depth_checked}
\title{An example sequence information file.}
\format{
Three lists with 201 factors, each contains 29 variables:
}
\usage{
mut_depth_checked
}
\description{
A dataset containing the information of reads for homology search.
}
\keyword{datasets}
